move uint 64 parser out
