+++ lib/core/crypto/eos_models/eos_name.dart
@@ -1,34 +1,6 @@
 import 'dart:typed_data';
 
-class Uint64Parser {
-  final int value;
-
-  Uint64Parser(this.value);
-
-  factory Uint64Parser.fromBytesBE(Uint8List bytes) {
-    final buffer = ByteData.view(bytes.buffer);
-    // final high = buffer.getUint32(0);
-    // final low = buffer.getUint32(4);
-
-    //return Uint64((high << 32) | low);
-    return Uint64Parser(buffer.getUint64(0));
-  }
-  factory Uint64Parser.fromBytesLE(Uint8List bytes) {
-    int value = 0;
-    for (int i = 7; i >= 0; i--) {
-      value = (value << 8) | bytes[i];
-    }
-    return Uint64Parser(value);
-  }
-
-  Uint8List toBytesLE() {
-    final bytes = Uint8List(8);
-    for (int i = 0; i < 8; i++) {
-      bytes[i] = (value >> (i * 8)) & 0xFF;
-    }
-    return bytes;
-  }
-}
+import 'package:hypha_wallet/core/crypto/eos_models/uint_64_parser.dart';
 
 class EosName {
   int value;
